+++ src/math/poly.h
@@ -13,7 +13,7 @@ namespace math {
  */
 void polyfit(const cv::Mat& src_x, const cv::Mat& src_y, cv::Mat& dst,
              int order);
-std::vector<double> polyfit(const std::vector<cv::Point>& src, int order);
+std::vector<float> polyfit(const std::vector<cv::Point>& src, int order);
 
 }  // namespace math
 }  // namespace robocar
